+++ src/edu/nyu/cs/cs2580/IndexerInvertedOccurrence.java
@@ -53,14 +53,17 @@ public class IndexerInvertedOccurrence extends Indexer implements Serializable{
 
     //Clearing all Index files
     File indexDir = new File(_options._indexPrefix);
+    indexDir.mkdir();
     File[] foundFiles = indexDir.listFiles(new FilenameFilter() {
       public boolean accept(File dir, String name) {
         return true;
       }
     });
 
-    for (File file : foundFiles) {
-      file.delete();
+    if(foundFiles != null) {
+      for (File file : foundFiles) {
+        file.delete();
+      }
     }
 
     try {
@@ -124,8 +127,10 @@ public class IndexerInvertedOccurrence extends Indexer implements Serializable{
         }
       });
 
-      for (File file : foundFiles) {
-        file.delete();
+      if(foundFiles != null) {
+        for (File file : foundFiles) {
+          file.delete();
+        }
       }
 
       long endTime = System.currentTimeMillis();
